requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
paho-mqtt==2.1.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.3.0
